requirements.txt
discord.py
python-dotenv~=0.15.0
